Create README.md
# olio

These are some "olio" practis task that we made in class. 
These "olio" tasks were verry interesting. 
We also did some practis with Rest apis but when i later needed to use this on my own website I noticed that it is really hard to find one that works and is what
you are looking for.